requirements.txt
streamlit
numpy
pillow
cryptography
